methods

+++ tests/test_10.java
@@ -5,7 +5,7 @@ public class test10 {
     int x = 4;
     public static void main(String[] args) {
         System.out.println("Hello World!");
-        float a=5, vv=x, zz=10;  
+        int a=5, vv=x, zz=10;  
         float z=x;
         float zzz = (1.1f + 2);
         assignmnet(a, a, a, a);
